Additional test updates.
