+++ src/Microsoft.Health.Fhir.Tests.Common/TestFiles/Stu3/Bundle-Batch.json
@@ -198,7 +198,7 @@
             "request": {
                 "method": "GET",
                 "url": "Patient/12334",
-                "ifNoneMatch": "W/\"4\"",
+                "ifNoneMatch": "W/\"4a\"",
                 "ifModifiedSince": "2015-08-31T08:14:33+10:00"
             }
         }
